Refactoring
- Move clear_screen() and print_progress_bar() to thepower.py
- Remove unused imports
